ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.0.0"
description = "AI image generation service — FastAPI + Stability AI + OpenAI + HuggingFace"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.1",
    "Pillow>=11.1.0",
    "openai>=1.58.1",
    "sentry-sdk[fastapi]>=2.19.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "boto3>=1.35.86",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
    "ruff>=0.8.6",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
